Median filter Attempt
This was an attempt to run a median filter on the data (scan by scan) and plot the results. It does not fix the issues we are seeing with many points emanating radially from the sensor's origin. The most it will do is cause corners to become more rounded and walls to become less "flat"